+++ semver-project-gradle-plugin/main/kotlin/com/javiersc/semver/project/gradle/plugin/tasks/WriteSemverTask.kt
@@ -56,8 +56,7 @@ constructor(
                     |$semverWithPrefix
                     |
                 """
-                    .trimMargin()
-            )
+                    .trimMargin())
         }
     }