+++ semver-project-gradle-plugin/test/kotlin/com/javiersc/semver/project/gradle/plugin/GitCommitTest.kt
@@ -56,8 +56,7 @@ internal class GitCommitTest {
             git.commitsInCurrentBranchRevCommit.map(RevCommit::getShortMessage).shouldBe(messages)
             git.commitsInCurrentBranch.map(GitRef.Commit::message).shouldBe(messages)
             git.commitsInCurrentBranchHash.shouldBe(
-                git.commitsInCurrentBranch.map(GitRef.Commit::hash)
-            )
+                git.commitsInCurrentBranch.map(GitRef.Commit::hash))
         }
     }